Move dragon directory into app directory with generation directory.
