+++ backend/app/dragon/table.js
@@ -1,7 +1,7 @@
 const pool = require("../../databasePool.js");
 
 class DragonTable {
-
+  
 }
 
 module.exports = DragonTable;
